Updated Teach Access Website
Updated the website to make it more accessable to color blind users.
I bolded every link on the website slides to help color impaired users
identify the the links and distinguish them from plaintext.  Please note
the main commits for these changes were made on the master branch and not
this branch - an accident.